# docs/swagger.yaml
swagger: "2.0"
info:
  description: "Полнофункциональный REST API для игры в покер с категориями столов LOW, MID, VIP. Использует PostgreSQL, Kafka, Redis и авторизацию через Telegram Web App"
  title: "Poker REST API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    name: "API Support"
    url: "http://www.swagger.io/support"
    email: "support@swagger.io"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
  version: "1.0"
host: "localhost:3000"
basePath: "/api/v1"
securityDefinitions:
  TelegramAuth:
    type: apiKey
    name: x-init-data
    in: header
    description: "Telegram Web App init_data для авторизации"
paths:
...
  /games/{gameId}:
    get:
      tags:
        - game
      summary: "Получить состояние игры"
      description: "Возвращает текущее состояние игры по ID"
      security:
        - TelegramAuth: []
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: gameId
          in: path
          type: string
          required: true
          description: "ID игры"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Game"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
  /games/{gameId}/action:
    post:
      tags:
        - game
      summary: "Сделать ход в игре"
      description: "Обрабатывает игровое действие игрока (fold, call, raise, check, bet)"
      security:
        - TelegramAuth: []
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: gameId
          in: path
          type: string
          required: true
          description: "ID игры"
        - name: action
          in: body
          required: true
          description: "Действие игрока"
          schema:
            type: object
            properties:
              action:
                type: string
                enum: [fold, call, raise, check, bet]
              amount:
                type: integer
      responses:
        200:
          description: "OK"
          schema:
            type: object
            properties:
              message:
                type: string
              game:
                $ref: "#/definitions/Game"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
definitions:
...
  Card:
    type: object
    properties:
      suit:
        type: string
      rank:
        type: string
      value:
        type: integer
  GamePlayer:
    type: object
    properties:
      id:
        type: integer
      game_id:
        type: string
      user_uuid:
        type: string
      position:
        type: integer
      cards:
        type: array
        items:
          $ref: "#/definitions/Card"
      chips:
        type: integer
      bet:
        type: integer
      is_folded:
        type: boolean
      is_all_in:
        type: boolean
      last_action:
        type: string
  Game:
    type: object
    properties:
      id:
        type: string
      table_id:
        type: integer
      state:
        type: string
        enum: [waiting, preflop, flop, turn, river, showdown, finished]
      community_cards:
        type: array
        items:
          $ref: "#/definitions/Card"
      pot:
        type: integer
      current_bet:
        type: integer
      dealer_position:
        type: integer
      current_player:
        type: integer
      small_blind:
        type: integer
      big_blind:
        type: integer
      players:
        type: array
        items:
          $ref: "#/definitions/GamePlayer"
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
  Error:
    type: object
    properties:
      error:
        type: string